Fix cohesive elements
The node ordering of the cohesive elements was sometimes wrong, as
documented in issue #14.  It was based on the information recorded in
struct BOUNDARY.BoundEltFacetOri, which was actually misleading, as it
was the orientation wrt the 2D element (in 3D) from which the boundary
element was created - not the 3D element it is attached to, as was
expected.  struct BOUNDARY.BoundEltFacetOri was actually of no use.

The struct BOUNDARY.BoundEltFacetOri member and the related code was
removed altogether.

To get a properly oriented boundary element (base normal pointing inside
the element, i.e. outside the first 3D element it is attached to, as
documented in issue #14), it is sufficient to use the nodes of the
facet of the first 3D element it is attached to (in their nominal
order).

Thanks @adsimap.  Fixes #14.