Fix argument reading
In cases where a file was loaded as argument, it could sometimes be
erroneously interprated as a multiscale cell file, leading to an error.
This is now fixed by using "msfile(...)" when loading a multiscale cell
file, instead of the standard "file(...)".  The standard "file()" will
still be supported even for a multiscale cell file except in the
problematic cases (backward compatibility).