Fix penalization in -morpho
Fixed computation of penalty (in the case of void cells).  Simple fix.

In some cases, and especially in 2D, problems occurred with optimization
being apparently trapped in a penalized state (very high objective
function).  This was simply due to the fact that the penalty was not
computed properly (reset).  Optimization now runs very smoothly.

Fixes #26.  Helps #25.