Change 2D 2nd-order cohesive elements
The 2D 2nd-order cohesive elements are now 6-node quadrangle elements
(with no intermediate nodes on the element edges linking the two parts)
instead of the standard 8-node quadrangle elements.

There is no such element in Gmsh.  The elements are given tag 100.
The mesh will not open in Gmsh.

There is no such element in Zset.  The elements are given label c2d4o2.
The mesh will not open in Zset.

I don't know if there is such an element in Abaqus.  The elements are
given label COH2D6.  I don't know whether the mesh will open in Abaqus.

Fixes #31 again.