Fix domain labels and add -transform resetlabels
- neut_tess_init_domain_label_3d tries to retrieve the actual face
labels in the case of a topologically cubic domain, but failed in
configurations where the domain was geometrically too far from a
well-aligned cube (some labels could be initialized several times and
other were not initialized at all).   This is now fixed by going back to
"f*" labels when this situation occurs.

- same change made to neut_tess_init_domain_label_2d.

- -transform resetlabels makes it possible to reset the labels of the
domain faces, edges and vertices of a tessellation.  This is a patch to
the previous issues.

Fixes rquey/neper#226.