Add preliminary support for multiphase FEPX sims
Removed error when a multiphase simulation with different numbers of
slip systems for the different phases is loaded.

This version is a preliminary fix and can be used for simulations for
which the slip system variables (slip, sliprate, and possibly others)
are not used.  Full support still need to be added.