Fix objective function of tesr
The 2/nv scaling factor has been moved to under the square root, which
makes the objective function independent of the number of voxels and
comparable to 1 (as it should be).  This change affects only the
objective function by a fixed scaling factor, which does not change the
result of the optimization.